fix dtype

+++ crit_utils.py
@@ -101,7 +101,7 @@ def lil_helper_boi(crit):
         start_time = crit.cells[0].rstart_time
         start_time = dt.strptime(start_time, '%Y-%m-%d_%H-%M-%S')
         age = start_time - birth
-        age = age + timedelta(hours = (crit.block_num * crit.hour_bins))
+        age = age + timedelta(hours = int((crit.block_num * crit.hour_bins)))
         geno = s.get_genotype(crit.animal)
         info = [crit.animal, crit.probe, crit.date, crit.time_frame, crit.block_num, birth, start_time, age, geno, crit.p_value_burst, crit.p_value_t, crit.dcc, (crit.p_value_burst > 0.05 and crit.p_value_t > 0.05), crit.kappa_burst, crit.kappa_t, crit.k2b, crit.k2t, crit.kprob_b, crit.kprob_t]
     except Exception as e:
